requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
actions-python-core==0.1.3
    # via -r requirements.in
aiofiles==24.1.0
    # via actions-python-core
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   gql
    #   httpx
    #   openai
attrs==24.2.0
    # via
    #   cattrs
    #   requests-cache
backoff==2.2.1
    # via gql
cattrs==23.2.3
    # via requests-cache
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   click-default-group
    #   llm
    #   sqlite-utils
click-default-group==1.2.4
    # via
    #   llm
    #   sqlite-utils
cryptography==43.0.0
    # via pyjwt
deprecated==1.2.14
    # via pygithub
distro==1.9.0
    # via openai
gql[requests]==3.5.0
    # via -r requirements.in
graphql-core==3.2.3
    # via gql
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   actions-python-core
    #   llm-openrouter
    #   openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.4
    # via -r requirements.in
jiter==0.5.0
    # via openai
llm==0.15
    # via
    #   -r requirements.in
    #   llm-openrouter
llm-openrouter==0.2
    # via -r requirements.in
markupsafe==2.1.5
    # via jinja2
multidict==6.0.5
    # via yarl
openai==1.40.3
    # via llm
platformdirs==4.2.2
    # via requests-cache
pluggy==1.5.0
    # via
    #   llm
    #   sqlite-utils
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   llm
    #   openai
pydantic-core==2.20.1
    # via pydantic
pygithub==2.3.0
    # via -r requirements.in
pyjwt[crypto]==2.9.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
python-dateutil==2.9.0.post0
    # via sqlite-utils
python-ulid==2.7.0
    # via llm
pyyaml==6.0.2
    # via llm
requests==2.32.3
    # via
    #   gql
    #   pygithub
    #   requests-cache
    #   requests-toolbelt
requests-cache==1.2.1
    # via -r requirements.in
requests-toolbelt==1.0.0
    # via gql
six==1.16.0
    # via
    #   python-dateutil
    #   url-normalize
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlite-fts4==1.0.3
    # via sqlite-utils
sqlite-migrate==0.1b0
    # via llm
sqlite-utils==3.37
    # via
    #   llm
    #   sqlite-migrate
tabulate==0.9.0
    # via sqlite-utils
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   pygithub
url-normalize==1.4.3
    # via requests-cache
urllib3==2.2.2
    # via
    #   pygithub
    #   requests
    #   requests-cache
wrapt==1.16.0
    # via deprecated
yarl==1.9.4
    # via gql

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
